[surfman] Xenstore check hang
That loop goes 'infinite' for XenFB devices if say dm-agent is restarted.
Also it is aimed to check that the domain info have disapeared from
Xenstore in case there is a GPU to FLR. This test seems to be done in
vgpu.c already (and should be done in the takedown callback anyway for
that specific purpose).