Add quirk for a common laptop resolution.
Unfortunately, many display adapters and drivers require that their modes have
a horizontal resolution which is divisible by 8. As a result, it's common to
see "rounded down" modes, which have a slight mismatch from the size of a panel.
A common case is to see a mode of 1360x768 displayed on 1366x768 panels.

In these cases, panel stretching isn't appropriate: it creates a slightly
stretched display without any significant gain. Instead, we'll center the
display on the panel.

Signed-off-by: Kyle J. Temkin <temkink@ainfosec.com>