[libsurfman] update pfn->mfn surfman array.
Surfman maintains a "p2m" array for the VRAM using an added hypercall to
do the translation (not recommended).
This should be mandatory with surface_update callbacks that might change
the list of pages scanned. For some unknown reason this is not done for
Xenfb PV surfaces (which uses a specific helper in libsurfman), so leave
it as it is now, but that might change later if required.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>