Fix DRM plugin code to both enumerate and monitor udev.
If Surfman loads the DRM plugin before the DRM device node that it is looking
for was created, the plugin must use a udev monitor to wait for it to show up.
If Surfman loads the DRM plugin after the DRM device node that it is looking
for was created, the plugin must enumerate current udev devices to find it.

OXT-634

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>