Round down resolutions that are not divisible by 8.
This is really a quirk following the logic in surfman PR #4 that does the same
checking. This change allows the "rounded down" xres to be reported to the
guest via the EDID as well as in the Vesa modes table (where e.g. 1366x768 was
changed to 1360x768). The value in the EDID is being used and set by XenVesaDO
when it first starts causing modeset failures when not rounded down.

OXT-640

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>