surfman: Load plugins for non-primary devices.
There is no guaranty on what device will be PCI_CLASS 0x300 or 0x380,
but the plugin should be loaded for this one anyway (plus restricting
to primary display was a hack for a defunct plugin).

It turns out to usually be the integrated one, also most firmware allows
that to be configured, but some platforms (e.g, HP 800 G1) will
__change__ the PCI class of the Intel device when a PCI(e) discrete card
is racked (0x300->0x380). Whatever happens next is unwanted.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-967