surfman: xc_domain_getinfo usage.
xc_domain_getinfo(h, domid, n, info):
will return the number of xc_dominfo_t written in info and fill
info starting with the first domain with domain_id >= domid.

In practice, this means, xc_domain_getinfo(h, domid, 1, info), may fill
the info struct for a domain which domain_id != domid. This is a problem
for xc_domid_exists() which makes that assumption.

Since libsurfman provides xc_domid_getinfo() handle the case in it and
use the exported function across surfman.

The main in this is xc_domid_getinfo() will return -ENOENT if the domain
associated with the domid in argument does not exist.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>