surfman: get_monitor_info() may fail.
Don't segv when that happens, just move on.
This can happen in a variety of scenari, either the plugin saw a new
monitor but something went bad and it can't fetch its info, or the
monitor was quickly plugged and yanked, so the plugin no longer has info
on it.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1653

(cherry picked from commit dcbde7400bb377dbec90c41e0aaac70e2ab3eb6a)
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>